finishing admin add, update and delete products
