+++ client/store/index.js
@@ -6,9 +6,8 @@ import user from './user';
 import products from './products';
 import cart from './cart';
 import users from './users';
-import adminstore from './adminstore';
 
-const reducer = combineReducers({user, products, cart, users, adminstore});
+const reducer = combineReducers({user, products, cart, users});
 
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
