+++ client/store/products.js
@@ -10,6 +10,9 @@ const initialState = {
 const SET_PRODUCTS = 'SET_PRODUCTS';
 const SET_PRODUCT = 'SET_PRODUCT';
 const UPDATED_QUANTITY = 'UPDATED_QUANTITY';
+const ADD_PRODUCT = 'ADD_PRODUCT';
+const UPDATE_PRODUCT = 'UPDATE_PRODUCT';
+const DELETE_PRODUCT = 'DELETE_PRODUCT';
 
 const setProducts = products => ({
   type: SET_PRODUCTS,
@@ -23,6 +26,21 @@ const setProduct = product => ({
 
 const updatedQuantity = product => ({
   type: UPDATED_QUANTITY,
+  product
+});
+
+const addProduct = product => ({
+  type: ADD_PRODUCT,
+  product
+});
+
+const updateProduct = product => ({
+  type: UPDATE_PRODUCT,
+  product
+});
+
+const deleteProduct = product => ({
+  type: DELETE_PRODUCT,
   product
 });
 
@@ -58,6 +76,39 @@ export const decreaseQuantity = productId => async dispatch => {
   dispatch(updatedQuantity(data));
 };
 
+export const addProductThunk = params => {
+  return async dispatch => {
+    try {
+      const {data} = await axios.post('/api/products', params);
+      dispatch(addProduct(data));
+    } catch (error) {
+      console.error(error);
+    }
+  };
+};
+
+export const updateProductThunk = (productId, params) => {
+  return async dispatch => {
+    try {
+      const {data} = await axios.put(`/api/products/${productId}`, params);
+      dispatch(updateProduct(data));
+    } catch (error) {
+      console.error(error);
+    }
+  };
+};
+
+export const deleteProductThunk = productId => {
+  return async dispatch => {
+    try {
+      const {data} = await axios.delete(`/api/products/${productId}`);
+      dispatch(deleteProduct(data));
+    } catch (error) {
+      console.error(error);
+    }
+  };
+};
+
 export default function(state = initialState, action) {
   switch (action.type) {
     case SET_PRODUCTS:
@@ -66,6 +117,17 @@ export default function(state = initialState, action) {
       return {...state, product: action.product, singleLoading: false};
     case UPDATED_QUANTITY:
       return {...state, product: action.product};
+    case ADD_PRODUCT:
+      return {...state, products: [...state.products, action.product]};
+    case UPDATE_PRODUCT:
+      return {...state, product: action.product};
+    case DELETE_PRODUCT:
+      return {
+        ...state,
+        products: state.products.filter(
+          product => product.id !== action.productId
+        )
+      };
     default:
       return state;
   }
